Compute Motif Energies

This function computes motif energies for given peak intervals using a specified motif database. Optionally, it can normalize the computed motif energies using additional normalization intervals.

Usage

compute_motif_energies(
  peak_intervals,
  db = iceqream::motif_db,
  normalization_intervals = peak_intervals,
  prior = 0.01,
  normalize = TRUE,
  energy_norm_quantile = 1,
  norm_energy_max = 10,
  min_energy = -7
)

Arguments

peak_intervals

A data frame containing the peak intervals with columns: chrom, start, and end. If an additional column 'peak_name' is present, it will be used as row names in the output matrix.

db

A motif database to use for extracting motif energies. Default is iceqream::motif_db.

normalization_intervals

A data frame containing intervals for normalization. Default is peak_intervals.

prior

a prior probability for each nucleotide.

normalize

A logical value indicating whether to normalize the motif energies. Default is TRUE.

energy_norm_quantile

A numeric value for the quantile used in normalization. Default is 1.

norm_energy_max

The maximum value to which the normalized energy values are scaled. Default is 10.

min_energy

The minimum energy value to be assigned after normalization. Default is -7.

Value

A matrix of motif energies.