Compute tracks quantiles

This function computes the quantiles for a set of ATAC-seq tracks.

Usage

compute_tracks_q(
  atac_names,
  atac_tracks,
  iterator = 20,
  norm_q = 0.995,
  norm_intervals = gintervals.all(),
  tn5bias_track = "tn5bias",
  normalize_tn5bias = TRUE,
  sample_n = 10000
)

Arguments

atac_names

A character vector specifying the names of the ATAC-seq tracks.

atac_tracks

A list of numeric vectors representing the ATAC-seq tracks.

iterator

An iterator object used for computing quantiles.

norm_q

A numeric vector specifying the quantiles to compute for each track. If a single value, the same quantile is used for all tracks.

norm_intervals

A numeric vector specifying the intervals for computing quantiles.

tn5bias_track

A numeric vector representing the TN5 bias track.

normalize_tn5bias

Logical, whether to normalize the TN5 bias track.

sample_n

A numeric value specifying the maximum number of data points to use for computing quantiles.

Value

A data frame with columns "type" and "q", where "type" represents the name of the ATAC-seq track and "q" represents the computed quantiles.