Create Dinucleotide Feature Group

This function creates an IQFeatureGroup object for all dinucleotide features and GC content.

Usage

create_dinuc_feature_group(
  traj_model,
  quantile = 0.95,
  size = NULL,
  dinucleotides = c("AA", "AC", "AG", "AT", "CA", "CC", "CG", "CT", "GA", "GC", "GG",
    "GT", "TA", "TC", "TG", "TT"),
  include_gc = TRUE
)

Arguments

traj_model

A trajectory model object.

quantile

The quantile to use for normalization (default is 0.95).

size

The size of the sequences to use for feature computation.

dinucleotides

A vector of dinucleotides to create features for (default is all possible dinucleotides).

include_gc

Logical indicating whether to include GC content feature (default is TRUE).

Value

An IQFeatureGroup object containing all dinucleotide features and GC content if included.