Learn 'prego' models for ATAC difference of a trajectory

Learn 'prego' models for ATAC difference of a trajectory

Usage

learn_traj_prego(
  peak_intervals,
  atac_diff,
  n_motifs,
  min_diff = 0.2,
  energy_norm_quantile = 1,
  norm_energy_max = 10,
  min_energy = -7,
  sample_for_kmers = FALSE,
  sample_fraction = 0.1,
  sequences = NULL,
  seed = NULL,
  peaks_size = 300,
  additional_features = NULL,
  norm_intervals = peak_intervals,
  ...
)

Arguments

peak_intervals

A data frame, indicating the genomic positions ('chrom', 'start', 'end') of each peak, with an additional column named "const" indicating whether the peak is constitutive and therefiore shouldn't be used in the regression. Optionally, a column named "cluster" can be added with indication of the cluster of each peak.

atac_diff

A numeric vector, indicating the ATAC difference of each peak

n_motifs

Number of motifs to learn. Should be at least 2

min_diff

Minimum ATAC difference to include a peak in the training

energy_norm_quantile

quantile of the energy used for normalization. Default: 1

norm_energy_max

maximum value of the normalized energy. Default: 10

min_energy

Minimum energy value after normalization (default: -7)

sample_for_kmers

Whether to sample kmers for training. Default: TRUE

sample_fraction

Fraction of peaks to sample for training. Default: 0.1#'

sequences

A character vector of sequences to learn the motifs on. If NULL, the sequences of the peaks are used.

seed

Random seed

peaks_size

size of the peaks to extract sequences from. Default: 300bp

additional_features

A matrix of additional features to filter out before learning the motifs (e.g. CpG content, dinucleotide content, etc.)

norm_intervals

A data frame, indicating the genomic positions ('chrom', 'start', 'end') of peaks used for energy normalization. If NULL, the function will use peak_intervals for normalization.

...

Additional arguments to be passed to prego::regress_pwm