Motif Database

A compilation of DNA motifs from various sources.

Usage

motif_db

Format

A data frame with 270340 rows and 9 columns:

motif

Character. The name or identifier of the motif.

pos

Integer. The position within the motif.

A

Numeric. The score for adenine at this position.

C

Numeric. The score for cytosine at this position.

G

Numeric. The score for guanine at this position.

T

Numeric. The score for thymine at this position.

dataset

Character. The source dataset of the motif.

motif_orig

Character. The name of the motif without the database prefix.

cluster

Character. For SCENIC motifs, the cluster to which the motif belongs.

References

HOMER:

Heinz S, Benner C, Spann N, Bertolino E et al. Simple Combinations of Lineage-Determining Transcription Factors Prime cis-Regulatory Elements Required for Macrophage and B Cell Identities. Mol Cell 2010 May 28;38(4):576-589. PMID: 20513432

JASPAR:

Castro-Mondragon JA, Riudavets-Puig R, Rauluseviciute I, Berhanu Lemma R, Turchi L, Blanc-Mathieu R, Lucas J, Boddie P, Khan A, Manosalva Pérez N, Fornes O, Leung TY, Aguirre A, Hammal F, Schmelter D, Baranasic D, Ballester B, Sandelin A, Lenhard B, Vandepoele K, Wasserman WW, Parcy F, and Mathelier A JASPAR 2022: the 9th release of the open-access database of transcription factor binding profiles Nucleic Acids Res. 2022 Jan 7;50(D1):D165-D173.; doi: 10.1093/nar/gkab1113

JOLMA:

Jolma, A., Yin, Y., Nitta, K. et al. DNA-dependent formation of transcription factor pairs alters their binding specificity. Nature 534, S15–S16 (2016). https://doi.org/10.1038/nature18912

HOCOMOCO:

Ivan V. Kulakovskiy; Ilya E. Vorontsov; Ivan S. Yevshin; Ruslan N. Sharipov; Alla D. Fedorova; Eugene I. Rumynskiy; Yulia A. Medvedeva; Arturo Magana-Mora; Vladimir B. Bajic; Dmitry A. Papatsenko; Fedor A. Kolpakov; Vsevolod J. Makeev: HOCOMOCO: towards a complete collection of transcription factor binding models for human and mouse via large-scale ChIP-Seq analysis. Nucl. Acids Res., Database issue, gkx1106 (11 November 2017). https://doi.org/10.1093/nar/gkx1106

SCENIC:

Aibar, S., González-Blas, C., Moerman, T. et al. SCENIC: single-cell regulatory network inference and clustering. Nat Methods 14, 1083–1086 (2017). https://doi.org/10.1038/nmeth.4463

SCENIC+:

Bravo González-Blas, C., De Winter, S., Hulselmans, G. et al. SCENIC+: single-cell multiomic inference of enhancers and gene regulatory networks. Nat Methods 20, 1355–1367 (2023). https://doi.org/10.1038/s41592-023-01938-4

Examples

data(motif_db)
head(motif_db)
#>                                       motif pos            A            C
#> 1 SCENIC.kznf__ZNF264_Imbeault2017_OM_RCADE   1 0.0009524922 0.0009544146
#> 2 SCENIC.kznf__ZNF264_Imbeault2017_OM_RCADE   2 0.0512315113 0.0812262143
#> 3 SCENIC.kznf__ZNF264_Imbeault2017_OM_RCADE   3 0.0010845939 0.6574809936
#> 4 SCENIC.kznf__ZNF264_Imbeault2017_OM_RCADE   4 0.2605162531 0.1952148060
#> 5 SCENIC.kznf__ZNF264_Imbeault2017_OM_RCADE   5 0.0009498837 0.5682242112
#> 6 SCENIC.kznf__ZNF264_Imbeault2017_OM_RCADE   6 0.0013589649 0.0776041520
#>            G          T dataset                         motif_orig cluster
#> 1 0.56711077 0.56534915  SCENIC kznf__ZNF264_Imbeault2017_OM_RCADE     0.1
#> 2 0.52173304 0.53077843  SCENIC kznf__ZNF264_Imbeault2017_OM_RCADE     0.1
#> 3 0.07061333 0.37197952  SCENIC kznf__ZNF264_Imbeault2017_OM_RCADE     0.1
#> 4 0.39681136 0.44779432  SCENIC kznf__ZNF264_Imbeault2017_OM_RCADE     0.1
#> 5 0.49618562 0.07882282  SCENIC kznf__ZNF264_Imbeault2017_OM_RCADE     0.1
#> 6 0.09271003 0.85316236  SCENIC kznf__ZNF264_Imbeault2017_OM_RCADE     0.1