Plot multi-trajectory model report

This function plots a report for a multi-trajectory model. The report includes the motifs, spatial models, and partial responses for each trajectory model when type is "pr". If type is "boxp", the function will plot the boxplots of the energy vs response for each motif. If type is "spat", the function will plot the spatial distribution of the motifs.

Usage

plot_multi_traj_model_report(
  multi_traj,
  filename = NULL,
  width = NULL,
  height = NULL,
  dev = grDevices::pdf,
  title = NULL,
  use_full = FALSE,
  names_map = NULL,
  type = "pr",
  ...
)

Arguments

multi_traj

A TrajectoryModelMulti object.

filename

Optional. The name of the output file. If not provided, the plot will be displayed on the screen.

width

Optional. The width of the output plot in inches.

height

Optional. The height of the output plot in inches.

dev

Optional. The graphics device to use for output. Default is 'pdf'.

title

Optional. The title of the plot.

use_full

Optional. Whether to use the models with all the motifs or not. Default is FALSE.

names_map

a named vector to map the names of the motifs to new names.

type

Optional. The type of the plot, can be "pr", "boxp" or "spat". Default is "pr".

...

Additional arguments to be passed to the plotting function.

Value

A ggplot2 object (invsibly if a filename is provided).