Convert a trajectory model to a list of PBM objects

This function takes a trajectory model and converts it into a list of PBM objects.

Usage

traj_model_to_pbm_list(
  traj_model,
  func = "logSumExp",
  normalization_intervals = traj_model@normalization_intervals,
  bits_threshold = NULL
)

Arguments

traj_model

A trajectory model object

func

The function to use for computing the energies "logSumExp" or "max". Default is "logSumExp"

normalization_intervals

The normalization intervals to use for computing the energies. Default is the normalization intervals of the trajectory model.

bits_threshold

The threshold for trimming the PSSM (default is NULL, no trimming)

Value

A list of PBM objects