Create Dinucleotide Features

This function creates IQSeqFeature objects for all possible dinucleotides.

Usage

create_dinuc_features(
  traj_model,
  quantile = 0.99,
  size = traj_model@params$peaks_size,
  dinucleotides = c("AA", "AC", "AG", "AT", "CA", "CC", "CG", "CT", "GA", "GC", "GG",
    "GT", "TA", "TC", "TG", "TT"),
  feat_names = dinucleotides
)

Arguments

traj_model

A trajectory model object.

quantile

The quantile to use for normalization (default is 0.99).

size

The size of the sequences to use for feature computation.

dinucleotides

A vector of dinucleotides to create features for (default is all possible dinucleotides).

feat_names

A vector of feature names to use for each dinucleotide (default is the dinucleotide itself).

Value

A list of IQSeqFeature objects, one for each dinucleotide.